interaction_service/requirements.txt
boto3
fastapi
uvicorn
passlib
python-jose
python-dotenv
PyJWT